Factor out duplicate code
