+++ index.d.mts
@@ -1,23 +1,5 @@
-export class InvalidTokenError extends Error {}
-
-export interface JwtDecodeOptions {
-  header?: boolean;
-}
-
-export interface JwtHeader {
-  typ?: string;
-  alg?: string;
-}
-
-export interface JwtPayload {
-  iss?: string;
-  sub?: string;
-  aud?: string[] | string;
-  exp?: number;
-  nbf?: number;
-  iat?: number;
-  jti?: string;
-}
+import { JwtDecodeOptions } from "./common";
+export * from "./common";
 
 declare function jwtDecode<T = unknown>(
   token: string,
